Fix adhernce value -> percentage
